chore(zk-explorer): Do not `'use client'` in Submit project page
