+++ apps/zk-explorer/app/submit/page.tsx
@@ -1,21 +1,9 @@
-'use client';
-
-import {
-  Button,
-  Input,
-  Card,
-  Typography,
-} from '@webb-tools/webb-ui-components';
+import { Card, Typography } from '@webb-tools/webb-ui-components';
 import { ArrowRightUp } from '@webb-tools/icons';
-import { useState } from 'react';
-import { handleSubmitProject, validateGithubUrl } from '../../utils/utils';
 import { LinkCard } from '../../components/LinkCard';
+import { SubmitPageControls } from '../../components/SubmitPageControls';
 
 export default function Submit() {
-  const [githubUrl, setGithubUrl] = useState('');
-  const [isValidGithubUrl, setIsValidGithubUrl] = useState(false);
-  const [errorMessage, setErrorMessage] = useState<string>();
-
   return (
     <main className="flex flex-col gap-6 pt-6">
       <Typography variant="h4" fw="bold">
@@ -26,48 +14,7 @@ export default function Submit() {
         <Card className="max-w-[780px]">
           <Typography variant="h5">GitHub Repository URL:</Typography>
 
-          <Input
-            autoFocus
-            id="repository url"
-            placeholder="https://github.com/username/repository-name"
-            errorMessage={errorMessage ? `*${errorMessage}` : undefined}
-            isInvalid={errorMessage !== undefined}
-            value={githubUrl}
-            onChange={(value) => {
-              setGithubUrl(value);
-
-              const isNewUrlValid = validateGithubUrl(value);
-
-              setIsValidGithubUrl(isNewUrlValid);
-
-              setErrorMessage(
-                value.length > 0 && !isNewUrlValid
-                  ? 'Invalid GitHub URL.'
-                  : undefined
-              );
-            }}
-          />
-
-          <div className="flex gap-4 flex-col sm:flex-row">
-            {/* TODO: Need link for this. */}
-            <Button href="#" variant="secondary" isFullWidth>
-              Learn More
-            </Button>
-
-            <Button
-              isFullWidth
-              isDisabled={!isValidGithubUrl}
-              onClick={async () => {
-                const errorMessage = await handleSubmitProject(githubUrl);
-
-                setErrorMessage(
-                  errorMessage !== null ? errorMessage : undefined
-                );
-              }}
-            >
-              Submit Project
-            </Button>
-          </div>
+          <SubmitPageControls />
         </Card>
 
         {/* TODO: Replace this with a link to the feedback form. */}
